refactor: flatten directory structure and use ResolutionStrategy enum directly
- Move actionHandler.ts to presentations folder root
- Move configRepositoryImpl.ts and gitRepositoryImpl.ts to infrastructures folder root
- Remove empty subdirectories (github-actions, config, git)
- Simplify strategy validation using ResolutionStrategy enum values directly
- Remove redundant strategy constants in favor of enum

This simplifies the project structure and reduces code duplication.
